chore: move functions
